Refactor move ClockElements into tick folder
