+++ src/app/tick/index.ts
@@ -1,0 +1,3 @@
+export { PercentElement } from './percent.element';
+export { ClockElement } from './clock.element';
+export { TickElement } from './tick.element';
